Database

Description

Documentation

initialiseDB :: IO Connection

Initialises the SQLite DB and creates table Games and Teams if not exists

saveFetchedGames :: Connection -> GameForParseArr -> IO ()

Saves games

queryAllGames :: Connection -> String -> IO [Game]

Gets all the games

queryAllTeams :: Connection -> IO [Team]

Gets all the teams

querySelectedGames :: Connection -> String -> IO [SpecificGame]

Fetches fullname and name for teams whose matches id of home teams in games